Leverage the dokuwiki cache to reduce requests
Instead of displaying all the time the data directly from planttext use
the handle method to gather the SVG code. The result of handle is cached
internally by dokuwiki, this way the server gather the SVG code, and
serve it.

This modification mean the server need to have access to internet, but
reduce greatly the number of requests done on planttext service.